refactored and extracted setup fixtures
