+++ tests/bdd/test_failed_scenario_modal.py
@@ -4,4 +4,4 @@ from tests.bdd.steps.common_steps import *
 from tests.bdd.steps.modal_steps import *
 from tests.bdd.steps.scenario_steps import *
 
-scenarios("features/scenario_modal.feature")
+scenarios("features/failed_scenario_modal.feature")
